#471 Relocate factory out of beta stages
