+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/core/boot/ApplicationManaged.java
@@ -15,7 +15,7 @@
  *  along with this library. If not, see {@literal<http://www.gnu.org/licenses/>}.
  */
 
-package org.dockbox.hartshorn.core.boot.beta;
+package org.dockbox.hartshorn.core.boot;
 
 public interface ApplicationManaged {
     ApplicationManager applicationManager();
