+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/core/boot/LifecycleObservable.java
@@ -15,10 +15,9 @@
  *  along with this library. If not, see {@literal<http://www.gnu.org/licenses/>}.
  */
 
-package org.dockbox.hartshorn.core.boot.beta;
+package org.dockbox.hartshorn.core.boot;
 
-import org.dockbox.hartshorn.core.context.ContextCarrier;
-
-public interface ApplicationManager extends ContextCarrier, ApplicationLogger, ApplicationProxier, LifecycleObservable {
-    boolean isCI();
+@FunctionalInterface
+public interface LifecycleObservable {
+    void register(LifecycleObserver observer);
 }
