+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/core/context/HartshornApplicationContext.java
@@ -42,10 +42,10 @@ import org.dockbox.hartshorn.core.binding.ContextWrappedHierarchy;
 import org.dockbox.hartshorn.core.binding.NativeBindingHierarchy;
 import org.dockbox.hartshorn.core.binding.Provider;
 import org.dockbox.hartshorn.core.binding.Providers;
-import org.dockbox.hartshorn.core.boot.beta.ApplicationLogger;
-import org.dockbox.hartshorn.core.boot.beta.ApplicationManager;
-import org.dockbox.hartshorn.core.boot.beta.ApplicationProxier;
-import org.dockbox.hartshorn.core.boot.beta.LifecycleObservable;
+import org.dockbox.hartshorn.core.boot.ApplicationLogger;
+import org.dockbox.hartshorn.core.boot.ApplicationManager;
+import org.dockbox.hartshorn.core.boot.ApplicationProxier;
+import org.dockbox.hartshorn.core.boot.LifecycleObservable;
 import org.dockbox.hartshorn.core.context.element.FieldContext;
 import org.dockbox.hartshorn.core.context.element.MethodContext;
 import org.dockbox.hartshorn.core.context.element.TypeContext;
