+++ hartshorn-persistence/src/main/java/org/dockbox/hartshorn/persistence/service/PersistentTypeService.java
@@ -18,7 +18,7 @@
 package org.dockbox.hartshorn.persistence.service;
 
 import org.dockbox.hartshorn.core.annotations.service.Service;
-import org.dockbox.hartshorn.core.boot.beta.LifecycleObserver;
+import org.dockbox.hartshorn.core.boot.LifecycleObserver;
 import org.dockbox.hartshorn.core.context.ApplicationContext;
 import org.dockbox.hartshorn.core.context.element.TypeContext;
 import org.dockbox.hartshorn.persistence.annotations.UsePersistence;
